requirements.txt
numpy
django
pillow
pathlib
opencv-python